pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devtoolhub"
version = "0.1.0"
description = "Centralized dashboard for local developer tools"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "textual>=1.0",
]

[project.scripts]
devtoolhub = "devtoolhub.__main__:main"

[tool.ruff]
line-length = 88
target-version = "py311"
